refactor(masumi): split the wild x402 dialect out of the payment-required schema
payment-required.schema.ts was 686 lines and the per-entry selection fix
would push it past the 750-line ceiling. Split by responsibility rather
than by line count:

- payment-required.wild.ts owns what Soko may READ — the lenient v1/v2
  dialect shapes, the per-field translations (maxAmountRequired -> amount,
  plain network name -> CAIP-2, address canonicalization) and the
  shadow-key filter.
- payment-required.supported.ts owns the option allowlists and the two
  free-form map shapes both sides need, so neither file has to import the
  other.
- payment-required.schema.ts keeps what Soko may EMIT: the node-shape
  schemas and the payload-level entry points.

Pure move. The one structural change is that the translation step now
hands back untyped data, which makes the trailing
x402PaymentRequiredSchema.safeParse the single gate on the emitted shape
instead of a second opinion on an already-typed value.
