+++ packages/masumi/src/schemas/index.ts
@@ -13,3 +13,4 @@ export * from "./input/validation.schema.js";
 // x402 schemas
 export * from "./x402/payment-required.limits.js";
 export * from "./x402/payment-required.schema.js";
+export * from "./x402/payment-required.supported.js";
